backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
bcrypt==4.2.0
PyJWT==2.9.0
pandas==2.2.3
openpyxl==3.1.5
pytest==8.3.3
httpx==0.27.2
pdfminer.six==20231228
google-generativeai==0.5.3


